move a few things around to different files
