+++ djstripe/models.py
@@ -20,14 +20,14 @@ from model_utils.models import TimeStampedModel
 import stripe
 
 from . import settings as djstripe_settings
+from djstripe.utils import convert_tstamp
 from .exceptions import SubscriptionCancellationFailure, SubscriptionUpdateFailure
 from .managers import CustomerManager, ChargeManager, TransferManager
 from .signals import WEBHOOK_SIGNALS
 from .signals import subscription_made, cancelled, card_changed
 from .signals import webhook_processing_error
 from . import webhooks
-from .stripe_objects import StripeEvent, StripeTransfer, StripeCustomer, StripeInvoice, StripeCharge, StripePlan, \
-    convert_tstamp
+from .stripe_objects import StripeEvent, StripeTransfer, StripeCustomer, StripeInvoice, StripeCharge, StripePlan
 
 logger = logging.getLogger(__name__)
