+++ djstripe/utils.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 import warnings
+import datetime
 
 from django.core.exceptions import ImproperlyConfigured
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser
-
-from .models import Customer
+from django.utils import timezone
+from django.conf import settings
 
 
 ANONYMOUS_USER_ERROR_MSG = (
@@ -43,6 +44,7 @@ def subscriber_has_active_subscription(subscriber):
     if isinstance(subscriber, get_user_model()):
         if subscriber.is_superuser or subscriber.is_staff:
             return True
+    from .models import Customer
 
     customer, created = Customer.get_or_create(subscriber)
     if created or not customer.has_active_subscription():
@@ -64,3 +66,36 @@ def get_supported_currency_choices(api_key):
 
     account = stripe.Account.retrieve()
     return [(currency, currency.upper()) for currency in account["currencies_supported"]]
+
+
+def dict_nested_accessor(d, name):
+    """
+    Access a dictionary value, possibly in a nested dictionary.
+    >>> dict_nested_accessor({'id': 'joe'}, 'id')
+    "joe"
+    >>> dict_nested_accessor({'inner': {'id': 'joe'}}, 'inner.id')
+    "joe"
+
+    :type d: dict
+    """
+    names = name.split(".", 1)
+    if len(names) > 1:
+        return dict_nested_accessor(d[names[0]], names[1])
+    else:
+        return d[name]
+
+
+def convert_tstamp(response, field_name=None):
+    """
+    Intended for use converting from a Stripe API timestamp resposne into a native date
+
+    :rtype: datetime
+    """
+    # Overrides the set timezone to UTC - I think...
+    tz = timezone.utc if settings.USE_TZ else None
+
+    if not field_name:
+        return datetime.datetime.fromtimestamp(response, tz)
+    else:
+        if field_name in response and response[field_name]:
+            return datetime.datetime.fromtimestamp(response[field_name], tz)
